Replace inheritance of NetworkX by encapsulation (#501)
* Step 1 - favour composition over inheritance.

* Step 2, expose interface and hide implementation - mostly.

* Removed explicit dependency on NetworkX.

* Use original NetworkX graph for NetworkX functionality.

* Fixed tests that were broken after merge from develop.

* Removed StellarGraph to NetworkX conversion method.

* Added assertion on special NetworkX test.

* Use set mutation.

* Reblacked code.

* Merged from develop; fixed cluster GCN.

* Moved NetworkX implementation to separate file.

* Fixed test script and a data URL.

* Renamed some methods.

* PyCharm missed one renaming.

* Updated changelog with a breaking change.
