+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 - Changed `GraphSAGE` and `HinSAGE` class API to accept generator objects the same as GCN/GAT models. Passing a `NodeSequence` or `LinkSequence` object is now deprecated.  [\#498](https://github.com/stellargraph/stellargraph/pull/498)
 
 **Breaking changes:**
+- The `StellarGraph` class no longer exposes `NetworkX` internals, only required functionality.
+In particular, calls like `list(G)` will no longer return a list of nodes; use `G.nodes()` instead.
+[\#297](https://github.com/stellargraph/stellargraph/issues/297)
+
 - Passing a `NodeSequence` or `LinkSequence` object to `GraphSAGE` and `HinSAGE` classes is now deprecated and no longer supported [\#498](https://github.com/stellargraph/stellargraph/pull/498).
 Users might need to update their calls of `GraphSAGE` and `HinSAGE` classes by passing `generator` objects instead of `generator.flow()` objects.
