+++ tests/data/test_converters.py
@@ -25,12 +25,13 @@ from stellargraph.data.converter import (
     NodeAttributeSpecification,
 )
 
+import networkx as nx
 import numpy as np
 import pytest
 
 
 def example_stellar_graph_1():
-    G = StellarGraph()
+    G = nx.Graph()
     elist = [(1, 2), (2, 3), (1, 4), (3, 2)]
 
     G.add_nodes_from([1, 2, 3, 4], label="")
@@ -41,7 +42,7 @@ def example_stellar_graph_1():
     G.nodes[3]["a1"] = 1
     G.nodes[1]["a2"] = 1
     G.nodes[4]["a2"] = 1
-    return G
+    return StellarGraph(G)
 
 
 def test_converter_categorical():
