+++ tests/mapper/test_link_mappers.py
@@ -436,7 +436,7 @@ class Test_GraphSAGELinkGenerator:
             feature_size=n_feat, n_nodes=6, n_isolates=2, n_edges=10
         )
 
-        unsupervisedSamples = UnsupervisedSampler(G, nodes=G.nodes)
+        unsupervisedSamples = UnsupervisedSampler(G, nodes=G.nodes())
 
         gen = GraphSAGELinkGenerator(G, batch_size=n_batch, num_samples=n_samples).flow(
             unsupervisedSamples
@@ -819,7 +819,7 @@ class Test_Attri2VecLinkGenerator:
             feature_size=n_feat, n_nodes=6, n_isolates=2, n_edges=10
         )
 
-        unsupervisedSamples = UnsupervisedSampler(G, nodes=G.nodes)
+        unsupervisedSamples = UnsupervisedSampler(G, nodes=G.nodes())
 
         gen = Attri2VecLinkGenerator(G, batch_size=n_batch).flow(unsupervisedSamples)
